feat: update ProxmoxExecutiveDashboard.php
