+++ app/Filament/Resources/ProxmoxClusterResource.php
@@ -18,7 +18,6 @@ use Filament\Tables\Actions\Action;
 use Filament\Tables\Actions\DeleteAction;
 use Filament\Tables\Actions\EditAction;
 use Filament\Tables\Columns\BadgeColumn;
-use Filament\Tables\Columns\ProgressColumn;
 use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Table;
 use Illuminate\Database\Eloquent\Builder;
@@ -156,15 +155,20 @@ class ProxmoxClusterResource extends Resource
                     ->badge()
                     ->color('success'),
 
-                ProgressColumn::make('health_score')
+                TextColumn::make('health_score')
                     ->label('Health')
-                    ->getStateUsing(function (ProxmoxCluster $record): int {
-                        return $record->getHealthScore();
+                    ->getStateUsing(function (ProxmoxCluster $record): string {
+                        $score = $record->getHealthScore();
+
+                        return $score.'%';
                     })
-                    ->color(function (int $state): string {
+                    ->badge()
+                    ->color(function (ProxmoxCluster $record): string {
+                        $score = $record->getHealthScore();
+
                         return match (true) {
-                            $state >= 80 => 'success',
-                            $state >= 60 => 'warning',
+                            $score >= 80 => 'success',
+                            $score >= 60 => 'warning',
                             default => 'danger',
                         };
                     }),
